// payload-types.ts
export interface Media {
  id: number;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    tablet?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Header {
  id: number;
  logo?: {
    text?: string | null;
    image?: (number | null) | Media;
  };
  navLinks?:
    | {
        label: string;
        /**
         * Internal link (e.g. /about) or external (e.g. https://example.com)
         */
        href: string;
        id?: string | null;
      }[]
    | null;
  ctaButton?: {
    text?: string | null;
    href?: string | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}